Represents discretized dense input.

Buckets include the left boundary, and exclude the right boundary. Namely, boundaries=[0., 1., 2.] generates buckets (-inf, 0.), [0., 1.), [1., 2.), and [2., +inf). For example, if the inputs are boundaries = [0, 10, 100] input tensor = [[-5, 10000] [150, 10] [5, 100]] then the output will be output = [[0, 3] [3, 2] [1, 3]]

[0., 1., 2.]: R:0.,
[0., 1.)`,
`[1., 2.)`, and `[2., +inf)`. For example, if the inputs are `boundaries` = [0, 10, 100]: R:0.,
[[-5, 10000]: R:[-5,
[150, 10]: R:150,
[5, 100]: R:5,
[[0, 3]: R:[0,
[3, 2]: R:3,
[1, 3]: R:1,

column_bucketized(source_column, boundaries)

Arguments

source_column
A one-dimensional dense column which is generated with numeric_column.

boundaries
A sorted list or list of floats specifying the boundaries.

Value

A _BucketizedColumn.

Raises

ValueError: If source_column is not a numeric column, or if it is not one-dimensional. ValueError: If boundaries is not a sorted list or list.

See also

Other feature_column wrappers: column_categorical_with_hash_bucket, column_categorical_with_identity, column_categorical_with_vocabulary_file, column_categorical_with_vocabulary_list, column_crossed, column_embedding, column_numeric, column_weighted_categorical