Represents sparse feature where ids are set by hashing.

Use this when your sparse features are in string or integer format, and you want to distribute your inputs into a finite number of buckets by hashing. output_id = Hash(input_feature_string) features, features$key$ is either Tensor or SparseTensor. If Tensor, missing values can be represented by -1 for int and '' for string. Note that these values are independent of the default_value argument.

column_categorical_with_hash_bucket(..., hash_bucket_size, dtype = tf$string)

Arguments

...
Expression(s) identifying input feature(s). Used as the column name and the dictionary key for feature parsing configs, feature Tensor objects, and feature columns.

hash_bucket_size
An int > 1. The number of buckets.

dtype
The type of features. Only string and integer types are supported.

Value

A _HashedCategoricalColumn.

Raises

ValueError: hash_bucket_size is not greater than 1. ValueError: dtype is neither string nor integer.

See also

Other feature_column wrappers: column_bucketized, column_categorical_with_identity, column_categorical_with_vocabulary_file, column_categorical_with_vocabulary_list, column_crossed, column_embedding, column_numeric, column_weighted_categorical, input_layer