Returns a column for performing crosses of categorical features.

Crossed features will be hashed according to hash_bucket_size. Conceptually, the transformation can be thought of as: Hash(cartesian product of features) SparseTensor referred by first key: shape = 2, 2 0, 0: "a" 1, 0: "b" 1, 1: "c" * SparseTensor referred by second key: shape = 2, 1 0, 0: "d" 1, 0: "e" then crossed feature will look like: shape = 2, 2 0, 0: Hash64("d", Hash64("a"))

column_crossed(keys, hash_bucket_size, hash_key = NULL)

Arguments

keys
An iterable identifying the features to be crossed. Each element can be either: * string: Will use the corresponding feature which must be of string type. * _CategoricalColumn: Will use the transformed tensor produced by this column. Does not support hashed categorical column.

hash_bucket_size
An int > 1. The number of buckets.

hash_key
Specify the hash_key that will be used by the FingerprintCat64 function to combine the crosses fingerprints on SparseCrossOp (optional).

Value

A _CrossedColumn.

Raises

ValueError: If len(keys) < 2. ValueError: If any of the keys is neither a string nor _CategoricalColumn. ValueError: If any of the keys is _HashedCategoricalColumn. ValueError: If hash_bucket_size < 1.

See also

Other feature_column wrappers: column_bucketized, column_categorical_with_hash_bucket, column_categorical_with_identity, column_categorical_with_vocabulary_file, column_categorical_with_vocabulary_list, column_embedding, column_numeric, column_weighted_categorical, input_layer