Dynamic RNN Estimator

Dynamic RNN Estimator

dynamic_rnn(problem_type, prediction_type, sequence_feature_columns,
  context_feature_columns = NULL, num_classes = NULL, num_units = NULL,
  cell_type = "basic_rnn", optimizer = "SGD", learning_rate = 0.1,
  predict_probabilities = F, momentum = NULL, gradient_clipping_norm = 5,
  dropout_keep_probabilities = NULL, feature_engineering_fn = NULL,
  model_dir = NULL, config = NULL)

Arguments

problem_type
whether the Estimator is intended for a regression or classification problem. Value must be one of ProblemType.CLASSIFICATION or ProblemType.LINEAR_REGRESSION.

prediction_type
whether the Estimator should return a value for each step in the sequence, or just a single value for the final time step. Must be one of PredictionType.SINGLE_VALUE or PredictionType.MULTIPLE_VALUE.

sequence_feature_columns
An iterable containing all the feature columns describing sequence features. All items in the iterable should be instances of classes derived from FeatureColumn.

context_feature_columns
An iterable containing all the feature columns describing context features, i.e., features that apply across all time steps. All items in the set should be instances of classes derived from FeatureColumn.

num_classes
the number of classes for a classification problem. Only used when problem_type=ProblemType.CLASSIFICATION.

num_units
A list of integers indicating the number of units in the RNNCells in each layer.

cell_type
A subclass of RNNCell or one of 'basic_rnn,' 'lstm' or 'gru'.

optimizer
Either the name of the optimizer to be used when training the model, or a tf$Optimizer instance. Defaults to the SGD optimizer.

learning_rate
Learning rate. This argument has no effect if optimizer is an instance of an Optimizer.

predict_probabilities
A boolean indicating whether to predict probabilities for all classes. Used only if problem_type is ProblemType.CLASSIFICATION

momentum
Momentum value. Only used if optimizer_type is 'Momentum'.

gradient_clipping_norm
Parameter used for gradient clipping. If NULL, then no clipping is performed.

dropout_keep_probabilities
a list of dropout probabilities or NULL. If a list is given, it must have length len(num_units) + 1. If NULL, then no dropout is applied.

feature_engineering_fn
Takes features and labels which are the output of input_fn and returns features and labels which will be fed into model_fn. Please check model_fn for a definition of features and labels.

model_dir
Directory to save the model parameters, graph, and so on. This can also be used to load checkpoints from the directory into a estimator to continue training a previously saved model.

config
A run configuration created by run_config(), used to configure the runtime settings.