Ops and Objects Returned From Model Function and Passed to Estimator

Estimator spec fully defines the model to be run by estimator().

estimator_spec(mode, predictions = NULL, loss = NULL, train_op = NULL,
  eval_metric_ops = NULL, export_outputs = NULL,
  training_chief_hooks = NULL, training_hooks = NULL, scaffold = NULL)

Arguments

mode
A key that specifies if this is training, evaluation or prediction by mode_keys()

predictions
Predictions tensor or list of tensors.

loss
Training loss tensor. Must be either scalar, or with shape c(1).

train_op
Op for the training step.

eval_metric_ops
List of metric results keyed by name. The values of the dict are the results of calling a metric function, namely a (metric_tensor, update_op) list.

export_outputs
Describes the output signatures to be exported to saved model and used during serving. A list(name = output) where:

name: An arbitrary name for this output.

output: an export output object such as classification output, regression output, or prediction output. Single-headed models only need to specify one entry in this dictionary. Multi-headed models should specify one entry for each head, one of which must be named using signature_constants.DEFAULT_SERVING_SIGNATURE_DEF_KEY.

training_chief_hooks
List of session run hooks to run on the chief worker during training.

training_hooks
List of session run hooks that to run on all workers during training.

scaffold
A tf$train$Scaffold object that can be used to set initialization, saver, and more to be used in training.

See also

Other custom estimator methods: coef.tf_estimator, estimator, evaluate.tf_estimator, export_savedmodel.tf_estimator, predict.tf_estimator, train.tf_estimator