Evaluates the model given evaluation data input_fn.

For each step, calls input_fn, which returns one batch of data. Evaluates until:

steps batches are processed, or

input_fn raises an end-of-input exception (OutOfRangeError or StopIteration).

# S3 method for tf_estimator
evaluate(object, input_fn, steps = NULL,
  checkpoint_path = NULL, name = NULL, hooks = NULL, ...)

Arguments

object
A TensorFlow estimator.

input_fn
Input function returning a list of: features - Dictionary of string feature name to Tensor or SparseTensor. labels - Tensor or dictionary of Tensor with labels.

steps
Number of steps for which to evaluate model. If NULL, evaluates until input_fn raises an end-of-input exception.

checkpoint_path
Path of a specific checkpoint to evaluate. If NULL, the latest checkpoint in model_dir is used.

name
Name of the evaluation if user needs to run multiple evaluations on different data sets, such as on training data vs test data. Metrics for different evaluations are saved in separate folders, and appear separately in tensorboard.

hooks
List of SessionRunHook subclass instances. Used for callbacks inside the evaluation call.

...
Optional arguments passed on to the estimator's evaluate() method.

Value

A dict containing the evaluation metrics specified in model_fn keyed by name, as well as an entry global_step which contains the value of the global step for which this evaluation was performed.

Raises

ValueError: If steps <= 0. ValueError: If no model has been trained, namely model_dir, or the given checkpoint_path is empty.

See also

Other custom estimator methods: coef.tf_estimator, estimator_spec, estimator, export_savedmodel.tf_estimator, predict.tf_estimator, train.tf_estimator