Evaluate on The Evaluation Data

Runs evaluation on the evaluation data and returns the result. Runs for self._eval_steps steps, or if it's NULL, then run until input is exhausted or another exception is raised. Start the evaluation after delay_secs seconds, or if it's NULL, defaults to using self._eval_delay_secs seconds.

# S3 method for tf_experiment
evaluate(object, delay_secs = NULL, ...)

Arguments

object
A TensorFlow experiment.

delay_secs
Start evaluating after this many seconds. If NULL, defaults to using self._eval_delays_secs.

...
Optional arguments passed on to the experiment's evaluate() method.

Value

The result of the evaluate call to the Estimator.

See also

Other experiment methods: experiment.tf_estimator, train.tf_experiment, train_and_evaluate.tf_experiment