Experiment constructor

An Experiment contains all information needed to train a model. After an experiment is created (by passing an Estimator and inputs for training and evaluation), an Experiment instance knows how to invoke training and eval loops in a sensible fashion for distributed training.

# S3 method for tf_model
experiment(object, train_input_fn, eval_input_fn,
  train_steps = 2L, eval_steps = 2L, eval_metrics = NULL,
  train_monitors = NULL, eval_hooks = NULL, local_eval_frequency = NULL,
  eval_delay_secs = 120L, continuous_eval_throttle_secs = 60L,
  min_eval_frequency = NULL, delay_workers_by_global_step = NULL,
  export_strategies = NULL, train_steps_per_iteration = NULL)

Arguments

train_input_fn
function, returns features and labels for training.

eval_input_fn
function, returns features and labels for evaluation. Ifeval_steps is None, this should be configured only to produce for a finite number of batches (generally, 1 epoch over the evaluation data).

train_steps
Perform this many steps of training. NULL, the default, means train forever.

eval_steps
evaluate runs until input is exhausted (or another exception is raised), or for eval_steps steps, if specified.

eval_metrics
list of string, metric function. If NULL, default set is used. This should be NULL if the estimator is tf.estimator.Estimator. If metrics are provided they will be appended to the default set.

train_monitors
A list of monitors to pass to the Estimator's fit function.

eval_hooks
A list of SessionRunHook hooks to pass to the Estimator's evaluate function.

local_eval_frequency
(applies only to local_run) Frequency of running eval in steps. If None, runs evaluation only at the end of training.

eval_delay_secs
Start evaluating after waiting for this many seconds.

continuous_eval_throttle_secs
Do not re-evaluate unless the last evaluation was started at least this many seconds ago for continuous_eval().

delay_workers_by_global_step
if TRUE delays training workers based on global step instead of time.

export_strategies
Iterable of ExportStrategys, or a single one, or NULL.

train_steps_per_iteration
(applies only to continuous_train_and_eval). Perform this many (integer) number of train steps for each training-evaluation iteration. With a small value, the model will be evaluated more frequently with more checkpoints saved. If NULL, will use a default value (which is smaller than train_steps if provided).

min_eval_frequency:
(applies only to train_and_evaluate). the minimum number of steps between evaluations. Of course, evaluation does not occur if no new snapshot is available, hence, this is the minimum. If 0, the evaluation will only happen after training. If NULL, defaults to 1, unless model_dir is on GCS, in which case the default is 1000.

See also

Other experiment methods: evaluate.tf_experiment, train.tf_experiment, train_and_evaluate.tf_experiment