Prints the given tensors every N local steps, every N seconds, or at end.

The tensors will be printed to the log, with INFO severity.

hook_logging_tensor(tensors, every_n_iter = NULL, every_n_secs = NULL,
  formatter = NULL, at_end = FALSE)

Arguments

tensors
dict that maps string-valued tags to tensors/tensor names, or iterable of tensors/tensor names.

every_n_iter
int, print the values of tensors once every N local steps taken on the current worker.

every_n_secs
int or float, print the values of tensors once every N seconds. Exactly one of every_n_iter and every_n_secs should be provided.

formatter
function, takes dict of tag->Tensor and returns a string. If NULL uses default printing all tensors.

at_end
bool specifying whether to print the values of tensors at the end of the run.

Details

Note that if at_end is True, tensors should not include any tensor whose evaluation produces a side effect such as consuming additional inputs.

See also

Other session_run_hook wrappers: EstimatorSessionRunHook, hook_checkpoint_saver, hook_global_step_waiter, hook_nan_tensor, hook_step_counter, hook_stop_at_step, hook_summary_saver