Construct input function from a data.frame object used to feed the estimator.

Construct input function from a data.frame object used to feed the estimator.

# S3 method for data.frame
input_fn(x, features, response = NULL,
  batch_size = 10L, shuffle = TRUE, num_epochs = 1L,
  queue_capacity = 1000L, num_threads = 1L)

Arguments

x
The data.frame that represents the input source

features
The names of features to be used

response
The response variable name to be used

batch_size
The size of batches

shuffle
Whether to shuffles the queue. Avoid shuffle at prediction time

num_epochs
The number of epochs to iterate over data. If NULL will run forever.

queue_capacity
The size of queue to accumulate.

num_threads
The number of threads used for reading and enqueueing. In order to have predicted and repeatable order of reading and enqueueing, such as in prediction and evaluation mode, num_threads should be 1.

See also

Other input function constructors: input_fn.default, input_fn.list, input_fn.matrix

Examples

features <- c("drat", "cyl")
input_fn1 <- input_fn(mtcars, response = "mpg", features = features)