Construct input function from a list object used to feed the estimator.

This is particularly useful when constructing dynamic length of inputs for models like recurrent neural networks.

# S3 method for list
input_fn(x, features, response)

Arguments

x
The list that represents the input source

features
The names of features to be used

response
The response variable name to be used

See also

Other input function constructors: input_fn.data.frame, input_fn.default, input_fn.matrix

Examples

input_fn(
   x = list(
     feature_names = list(
       list(list(1), list(2), list(3)),
       list(list(4), list(5), list(6))),
     response = list(
       list(1, 2, 3), list(4, 5, 6))),
   features = c("feature_names"),
   response = "response")

#> function(features_as_named_list = T) {
#>     if (features_as_named_list) {
#>       inputs <- tf$constant(
#>         np$array(
#>           x$features,
#>           dtype = np$int64
#>         )
#>       )
#>       labels <- tf$constant(
#>         np$array(
#>           x$response
#>         )
#>       )
#>     } else {
#>       stop("input_fn.list() does not support custom estimator yet")
#>     }
#>     list(list(inputs = inputs), labels)
#>   }
#> <environment: 0x10d506a38>