Returns a dense Tensor as input layer based on given feature_columns.

Generally a single example in training data is described with FeatureColumns. At the first layer of the model, this column oriented data should be converted to a single Tensor. Example:

input_layer(features, feature_columns, weight_collections = NULL,
  trainable = TRUE)

Arguments

features
A mapping from key to tensors. _FeatureColumns look up via these keys. For example numeric_column('price') will look at 'price' key in this dict. Values can be a SparseTensor or a Tensor depends on corresponding _FeatureColumn.

feature_columns
An iterable containing the FeatureColumns to use as inputs to your model. All items should be instances of classes derived from _DenseColumn such as numeric_column, embedding_column, bucketized_column, indicator_column. If you have categorical features, you can wrap them with an embedding_column or indicator_column.

weight_collections
A list of collection names to which the Variable will be added. Note that, variables will also be added to collections tf.GraphKeys.GLOBAL_VARIABLES and ops.GraphKeys.MODEL_VARIABLES.

trainable
If TRUE also add the variable to the graph collection GraphKeys.TRAINABLE_VARIABLES (see tf.Variable).

Value

A Tensor which represents input layer of a model. Its shape is (batch_size, first_layer_dimension) and its dtype is float32. first_layer_dimension is determined based on given feature_columns.

Details

price <- column_numeric('price')
keywords_embedded <- column_embedding(categorical_column_with_hash_bucket("keywords", 1000), dimensions = 16)
columns <- c(price, keywords_embedded, ...)
features <- tf$parse_example(..., features = make_parse_example_spec(columns))
dense_tensor <- input_layer(features, columns)
prediction <- tf$layers$dense(dense_tensor, 1)

Raises

ValueError: if an item in feature_columns is not a _DenseColumn.

See also

Other feature_column wrappers: column_bucketized, column_categorical_with_hash_bucket, column_categorical_with_identity, column_categorical_with_vocabulary_file, column_categorical_with_vocabulary_list, column_crossed, column_embedding, column_numeric, column_weighted_categorical