Linear regressor model.

Train a linear regression model to predict label value given observation of feature values.

linear_regressor(feature_columns, model_dir = NULL,
  weight_column_name = NULL, optimizer = NULL, gradient_clip_norm = NULL,
  enable_centered_bias = FALSE, label_dimension = 1L, config = NULL,
  feature_engineering_fn = NULL)

Arguments

feature_columns
An iterable containing all the feature columns used by the model. All items in the set should be instances of classes derived from FeatureColumn.

model_dir
Directory to save model parameters, graph, etc. This can also be used to load checkpoints from the directory into a estimator to continue training a previously saved model.

weight_column_name
A string defining feature column name representing weights. It is used to down weight or boost examples during training. It will be multiplied by the loss of the example.

optimizer
An instance of tf.Optimizer used to train the model. If NULL, will use an Ftrl optimizer.

gradient_clip_norm
A float > 0. If provided, gradients are clipped to their global norm with this clipping ratio. See tf.clip_by_global_norm for more details.

enable_centered_bias
A bool. If TRUE, estimator will learn a centered bias variable for each class. Rest of the model structure learns the residual after centered bias.

label_dimension
Number of regression targets per example. This is the size of the last dimension of the labels and logits Tensor objects (typically, these have shape [batch_size, label_dimension]).

config
RunConfig object to configure the runtime settings.

feature_engineering_fn
Feature engineering function. Takes features and labels which are the output of input_fn and returns features and labels which will be fed into the model.

Details

Input of fit and evaluate should have following features, otherwise there will be a KeyError:

if weight_column_name is not NULL: key=weight_column_name, value=a Tensor

for column in feature_columns:

if isinstance(column, SparseColumn): key=column.name, value=a SparseTensor

if isinstance(column, WeightedSparseColumn): key=id column name, value=a SparseTensor, key=weight column name, value=a SparseTensor

if isinstance(column, RealValuedColumn): key=column.name, value=a Tensor

See also

Other canned estimators: dynamic_rnn_estimator, linear_classifier, linear_dnn_combined_classifier, linear_dnn_combined_regressor, state_saving_rnn_estimator