Construct Input Function Containing Python Dictionaries of Numpy Arrays

This returns a function outputting features and target based on the dict of numpy arrays. The dict features has the same keys as the x.

numpy_input_fn(x, y = NULL, batch_size = 128, num_epochs = 1,
  shuffle = NULL, queue_capacity = 1000, num_threads = 1)

Arguments

x
dict of numpy array object.

y
numpy array object. NULL if absent.

batch_size
Integer, size of batches to return.

num_epochs
Integer, number of epochs to iterate over data. If NULL will run forever.

shuffle
Boolean, if TRUE shuffles the queue. Avoid shuffle at prediction time.

queue_capacity
Integer, size of queue to accumulate.

num_threads
Integer, number of threads used for reading and enqueueing. In order to have predicted and repeatable order of reading and enqueueing, such as in prediction and evaluation mode, num_threads should be 1. #'

Details

Note that this function is still experimental and should only be used if necessary, e.g. feed in data that's dictionary of numpy arrays.

Raises

ValueError: if the shape of y mismatches the shape of values in x (i.e., values in x have same shape). TypeError: x is not a dict or shuffle is not bool.

See also

Other input functions: input_fn