Returns predictions for given features.

Returns predictions for given features.

# S3 method for tf_custom_model
predict(object, input_fn, checkpoint_path = NULL,
  predict_keys = NULL, hooks = NULL, as_iterable = F)

Arguments

input_fn
Input function returning features which is a dictionary of string feature name to Tensor or SparseTensor. If it returns a list, first item is extracted as features. Prediction continues until input_fn raises an end-of-input exception (OutOfRangeError or StopIteration).

checkpoint_path
Path of a specific checkpoint to predict. If NULL, the latest checkpoint in model_dir is used.

predict_keys
list of str, name of the keys to predict. It is used if the EstimatorSpec.predictions is a dict. If predict_keys is used then rest of the predictions will be filtered from the dictionary. If NULL, returns all.

hooks
List of SessionRunHook subclass instances. Used for callbacks inside the prediction call.

Yields

Evaluated values of predictions tensors.

Raises

ValueError: Could not find a trained model in model_dir. ValueError: if batch length of predictions are not same. ValueError: If there is a conflict between predict_keys and predictions. For example if predict_keys is not NULL but EstimatorSpec.predictions is not a dict.

See also

Other custom estimator methods: estimator_spec, estimator, evaluate.tf_custom_model, export_savedmodel.tf_custom_model, fit.tf_custom_model