Interleaves training and evaluation.

The frequency of evaluation is controlled by the contructor arg min_eval_frequency. When this parameter is 0, evaluation happens only after training has completed. Note that evaluation cannot happen more frequently than checkpoints are taken. If no new snapshots are available when evaluation is supposed to occur, then evaluation doesn't happen for another min_eval_frequency steps (assuming a checkpoint is available at that point). Thus, settings min_eval_frequency to 1 means that the model will be evaluated everytime there is a new checkpoint. This is particular useful for a "Master" task in the cloud, whose responsibility it is to take checkpoints, evaluate those checkpoints, and write out summaries. Participating in training as the supervisor allows such a task to accomplish the first and last items, while performing evaluation allows for the second. Returns: The result of the evaluate call to the Estimator as well as the export results using the specified ExportStrategy.

# S3 method for tf_experiment
train_and_evaluate(object)

Value

The result of the evaluate call to the Estimator as well as the export results using the specified ExportStrategy.

See also

Other experiment methods: evaluate.tf_experiment, experiment.tf_model, train.tf_experiment