Ops and objects returned from a model_fn and passed to Estimator.

EstimatorSpec fully defines the model to be run by Estimator.

estimator_spec(mode, predictions = NULL, loss = NULL, train_op = NULL,
  eval_metric_ops = NULL, export_outputs = NULL,
  training_chief_hooks = NULL, training_hooks = NULL, scaffold = NULL)

Arguments

mode
A ModeKeys. Specifies if this is training, evaluation or prediction.

predictions
Predictions Tensor or dict of Tensor.

loss
Training loss Tensor. Must be either scalar, or with shape c(1).

train_op
Op for the training step.

eval_metric_ops
Dict of metric results keyed by name. The values of the dict are the results of calling a metric function, namely a (metric_tensor, update_op) list.

export_outputs
Describes the output signatures to be exported to SavedModel and used during serving. A dict {name: output} where:

name: An arbitrary name for this output.

output: an ExportOutput object such as ClassificationOutput, RegressionOutput, or PredictOutput. Single-headed models only need to specify one entry in this dictionary. Multi-headed models should specify one entry for each head, one of which must be named using signature_constants.DEFAULT_SERVING_SIGNATURE_DEF_KEY.

training_chief_hooks
Iterable of tf.train.SessionRunHook objects to run on the chief worker during training.

training_hooks
Iterable of tf.train.SessionRunHook objects that to run on all workers during training.

scaffold
A tf.train.Scaffold object that can be used to set initialization, saver, and more to be used in training.

See also

Other custom estimator methods: coef.tf_custom_model, estimator, evaluate.tf_custom_model, export_savedmodel.tf_custom_model, predict.tf_custom_model, train.tf_custom_model