Linear classifier model.

Train a linear model to classify instances into one of multiple possible classes. When number of possible classes is 2, this is binary classification.

linear_classifier(feature_columns, model_dir = NULL, n_classes = 2L,
  weight_column_name = NULL, optimizer = NULL, gradient_clip_norm = NULL,
  enable_centered_bias = FALSE, config = NULL,
  feature_engineering_fn = NULL)

Arguments

feature_columns
An iterable containing all the feature columns used by the model. All items in the set should be instances of classes derived from FeatureColumn.

model_dir
Directory to save model parameters, graph and etc. This can also be used to load checkpoints from the directory into a estimator to continue training a previously saved model.

n_classes
number of label classes. Default is binary classification. Note that class labels are integers representing the class index (i.e. values from 0 to n_classes-1). For arbitrary label values (e.g. string labels), convert to class indices first.

weight_column_name
A string defining feature column name representing weights. It is used to down weight or boost examples during training. It will be multiplied by the loss of the example.

optimizer
The optimizer used to train the model. If specified, it should be either an instance of tf.Optimizer or the SDCAOptimizer. If NULL, the Ftrl optimizer will be used.

gradient_clip_norm
A float > 0. If provided, gradients are clipped to their global norm with this clipping ratio. See tf.clip_by_global_norm for more details.

enable_centered_bias
A bool. If TRUE, estimator will learn a centered bias variable for each class. Rest of the model structure learns the residual after centered bias.

config
RunConfig object to configure the runtime settings.

feature_engineering_fn
Feature engineering function. Takes features and labels which are the output of input_fn and returns features and labels which will be fed into the model.

Details

Input of fit and evaluate should have following features, otherwise there will be a KeyError:

if weight_column_name is not NULL, a feature with key=weight_column_name whose value is a Tensor.

for each column in feature_columns:

if column is a SparseColumn, a feature with key=column.name whose value is a SparseTensor.

if column is a WeightedSparseColumn, two features: the first with key the id column name, the second with key the weight column name. Both features' value must be a SparseTensor.

if column is a RealValuedColumn, a feature with key=column.name whose value is a Tensor.

See also

Other canned estimators: dynamic_rnn, linear_dnn_combined_classifier, linear_dnn_combined_regressor, linear_regressor, state_saving_rnn